requirements.txt
flask
paho-mqtt
requests
streamlit
pandas
streamlit-autorefresh